move lookup/phrase operations into their own modules, require auth for anything in them (return a 401 otherwise)
