+++ public-src/javascripts/phraseStore.js
@@ -3,7 +3,7 @@ import {writable} from 'svelte/store';
 const phraseStore = writable(null);
 
 async function action(url) {
-	const phrases = await fetch(url)
+	const phrases = await fetch(`/phrases/${url}`)
 		.then(res => res.json());
 
 	phraseStore.set(phrases);
